requirements.txt
streamlit
langchain
langchain-groq
langchain-google-genai
langgraph
python-dotenv
pymupdf
python-docx
pandas